refactor: split default designer
In this commit, we split the DefaultDesigner component into Network
and Simulation tabs. We also added the default look for the
SimulationTab.
