+++ src/i18n/locales/en-US.json
@@ -112,11 +112,15 @@
   "cmps.designer.bitcoind.actions.SendOnChainModal.balanceError": "Amount must be less than the {{backendName}} balance of {{balance}} BTC",
   "cmps.designer.bitcoind.actions.SendOnChainModal.successTitle": "Successfully sent {{amount}} BTC",
   "cmps.designer.bitcoind.actions.SendOnChainModal.successDesc": "Transaction ID: {{txid}}",
-  "cmps.designer.default.DefaultSidebar.title": "Network Designer",
-  "cmps.designer.default.DefaultSidebar.mainDesc": "Click on an element in the designer to see details",
-  "cmps.designer.default.DefaultSidebar.addNodesTitle": "Add Nodes",
-  "cmps.designer.default.DefaultSidebar.addNodesDesc": "Drag a node below onto the canvas to add it to the network",
-  "cmps.designer.default.DefaultSidebar.checkUpdates": "Check for new Node Versions",
+  "cmps.designer.default.DefaultSidebar.networkTabTitle": "Network Designer",
+  "cmps.designer.default.NetworkDesignerTab.mainDesc": "Click on an element in the designer to see details",
+  "cmps.designer.default.NetworkDesignerTab.addNodesTitle": "Add Nodes",
+  "cmps.designer.default.NetworkDesignerTab.addNodesDesc": "Drag a node below onto the canvas to add it to the network",
+  "cmps.designer.default.NetworkDesignerTab.checkUpdates": "Check for new Node Versions",
+  "cmps.designer.default.DefaultSidebar.simulationTabTitle": "Simulation Designer",
+  "cmps.designer.default.SimulationDesignerTab.createBtn": "Add a new Simulation",
+  "cmps.designer.default.SimulationDesignerTab.title": "Lightning Network Simulation",
+  "cmps.designer.default.SimulationDesignerTab.emptyMsg": "No simulations created yet",
   "cmps.designer.default.ImageUpdatesModal.title": "Check for new Node Versions",
   "cmps.designer.default.ImageUpdatesModal.updatesTitle": "There are new node versions available",
   "cmps.designer.default.ImageUpdatesModal.updatesDesc": "Add the new versions to begin using them in your Lightning networks.",
